+++ gmc_database_api_server/app/api/yby654.go
@@ -15,42 +15,37 @@ import (
 
 func Get_Cluster(c echo.Context) (err error) {
 	// search_val := c.QueryParam("name")
-	clusterName := c.QueryParam("cluster")
-	KubernetesNODE, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/nodes", false)
-	clusterModel := GetCluster2(c)
-	k := gjson.Get(KubernetesNODE, "items").Array()
-	label := make(map[string]string)
-	annotation := make(map[string]string)
-	clusterRole := ""
-	for t, _ := range k {
-		labels := k[t].Get("metadata.labels")
-		err_label := json.Unmarshal([]byte(labels.String()), &label)
-		if err_label != nil {
-			fmt.Printf("Error : %s\n", err_label)
+	// clusterName := c.Param("name")
+	KubernetesNODE, err := common.GetModel(c, "nodes")
+	if common.FilterStr(KubernetesNODE, "status") == "Failure" {
+		// errReturn := KubernetesNS.Array()
+		errJson := make(map[string]string)
+		err_ns := json.Unmarshal([]byte(KubernetesNODE), &errJson)
+		if err_ns != nil {
+			fmt.Printf("err_ns : %s\n", err_ns)
 		}
-		for key, _ := range label {
-			if key == "node-role.kubernetes.io/master" {
-				clusterRole = "master"
-			}
-		}
-		clusterModel.Lable = label
-		annotations := k[t].Get("metadata.annotations")
-		err_annotation := json.Unmarshal([]byte(annotations.String()), &annotation)
-		if err_annotation != nil {
-			fmt.Printf("Error : %s\n", err_annotation)
-		}
-		clusterModel.Annotation = annotation
-		clusterModel.Name = k[t].Get("metadata.name").String()
-		createTime := k[t].Get("metadata.creationTimestamp").Time()
-		clusterModel.Version = k[t].Get("status.nodeInfo.kubeletVersion").String()
-		clusterModel.Os = k[t].Get("status.nodeInfo.operatingSystem").String() + " / " + k[t].Get("status.nodeInfo.osImage").String()
-		clusterModel.Kernel = k[t].Get("status.nodeInfo.kernelVersion").String()
-		clusterModel.Role = clusterRole
-		clusterModel.CreateAt = createTime
+		// 	common.ErrorMsg(c, http.StatusNotFound, err)
+		return c.JSON(http.StatusNotFound, errJson)
 	}
+	// KubernetesNODE2, err := common.GetModel2(c, "nodes")
+
+	fmt.Printf("[#]KubernetesNODE : %+v\n", KubernetesNODE)
+	clusterModel := GetCluster2(c)
+	clusterModel.Lable = common.Filter(KubernetesNODE, "metadata.labels")
+	clusterModel.Annotation = common.Filter(KubernetesNODE, "metadata.annotations")
+	clusterModel.Kernel = common.Finding(KubernetesNODE, "status.nodeInfo", "kernelVersion")
+	clusterModel.Os = common.Finding(KubernetesNODE, "status.nodeInfo", "operatingSystem") + " / " + common.Finding(KubernetesNODE, "status.nodeInfo", "osImage")
+	clusterModel.Version = common.Finding(KubernetesNODE, "status.nodeInfo", "kubeProxyVersion")
+	clusterModel.Gpu = ""
+	clusterModel.Network = ""
+	createTime := common.FilterStr(KubernetesNODE, "metadata.creationTimestamp")
+
+	timer, err := time.Parse(time.RFC3339, createTime)
+	clusterModel.CreateAt = timer
 
 	return c.JSON(http.StatusOK, echo.Map{"data": clusterModel})
 }
+
 func Get_http(c echo.Context) (err error) {
 
 	data, err := common.GetModel(c, "namespaces")
@@ -65,102 +60,36 @@ func Get_http(c echo.Context) (err error) {
 
 }
 func Get_Project(c echo.Context) (err error) {
-	search_val := c.QueryParam("name")
+	search_val := c.Param("name")
 	clusterName := c.QueryParam("cluster")
-	KubernetesNS, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/namespaces/"+search_val, false)
-	ResourceCount, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/"+search_val+"/application_resource", false)
 	projectModel := GetProject2(c)
-	var ResourceModel model.PROJECT_RESOURCE
+	KubernetesNS, err := common.GetModel(c, "namespaces")
+	if common.FilterStr(KubernetesNS, "status.phase") != "Active" {
+		// errReturn := KubernetesNS.Array()
+		errJson := make(map[string]string)
+		err_ns := json.Unmarshal([]byte(KubernetesNS), &errJson)
+		if err_ns != nil {
+			fmt.Printf("err_ns : %s\n", err_ns)
+		}
+		// 	common.ErrorMsg(c, http.StatusNotFound, err)
+		return c.JSON(http.StatusNotFound, errJson)
+	}
+	fmt.Printf("[#] status : %+v\n", common.FilterStr(KubernetesNS, "status.phase"))
 	fmt.Printf("[#]KubernetesNS : %+v\n", KubernetesNS)
-	if gjson.Get(KubernetesNS, "status").String() == "Failure" {
-		common.ErrorMsg(c, http.StatusOK, common.ErrNoData)
-		return
-	}
-	fmt.Printf("[#]ResourceCount : %+v\n", gjson.Get(ResourceCount, "count.pod_count"))
-	label := make(map[string]string)
-	annotation := make(map[string]string)
-	labels := gjson.Get(KubernetesNS, "metadata.labels")
-	err_label := json.Unmarshal([]byte(labels.String()), &label)
-	if err_label != nil {
-		fmt.Printf("err_label : %s\n", err_label)
-	}
-	projectModel.Lable = label
-	annotations := gjson.Get(KubernetesNS, "metadata.annotations")
-	err_annotation := json.Unmarshal([]byte(annotations.String()), &annotation)
-	if err_annotation != nil {
-		fmt.Printf("err_annotation : %s\n", err_annotation)
-	}
-	projectModel.Annotation = annotation
-	projectModel.Name = gjson.Get(KubernetesNS, "metadata.name").String()
-	projectModel.Status = gjson.Get(KubernetesNS, "status.phase").String()
-	createTime := gjson.Get(KubernetesNS, "metadata.creationTimestamp").Time()
-	projectModel.Created_at = createTime
+	projectModel.Lable = common.Filter(KubernetesNS, "metadata.labels")
+	projectModel.Annotation = common.Filter(KubernetesNS, "metadata.annotations")
 	projectModel.ClusterName = clusterName
-
+	// KubernetesNS, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/namespaces/"+search_val, false)
+	ResourceCount, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/"+search_val+"/application_resource", false)
+	var ResourceModel model.PROJECT_RESOURCE
 	ResourceModel.PodCount = StringToInt(gjson.Get(ResourceCount, "count.pod_count").String())
 	ResourceModel.DeploymentCount = StringToInt(gjson.Get(ResourceCount, "count.deployment_count").String())
 	ResourceModel.JobCount = StringToInt(gjson.Get(ResourceCount, "count.job_count").String())
 	ResourceModel.CronjobCount = StringToInt(gjson.Get(ResourceCount, "count.cronjob_count").String())
 	ResourceModel.ServiceCount = StringToInt(gjson.Get(ResourceCount, "count.service_count").String())
 	projectModel.Resource = ResourceModel
-	return c.JSON(http.StatusOK, echo.Map{"data": projectModel})
-}
-func Get_Deployment(c echo.Context) (err error) {
-	namespaceName := c.QueryParam("namespace")
-	clusterName := c.QueryParam("cluster")
-	workspaceName := c.QueryParam("workspace")
-	// KubernetesDEPLOY, _ := HttpRequest(c, "https://g-api-test.innogrid.tech/kube/v1/"+clusterName+"/"+namespaceName+"/deployments/"+search_val, false)
-	KubernetesDEPLOY, err := common.GetModel(c, "deployments")
-	var DeploymentModel model.DEPLOYMENT
-	var STRATEGY model.STRATEGY
-	var REPLICA model.REPLICA
-	// var CONTAINER model.CONTAINER
-	// label := make(map[string]string)
-	// annotation := make(map[string]string)
-	// fmt.Printf("[#]KubernetesDEPLOY : %+v\n", KubernetesDEPLOY)
-	DeploymentModel.Lable = common.Filter(KubernetesDEPLOY, "metadata.labels")
-	// DeploymentModel.Lable = labels
-	DeploymentModel.Annotation = common.Filter(KubernetesDEPLOY, "metadata.labels")
-	//  = annotaitions
-	createTime := common.FilterStr(KubernetesDEPLOY, "metadata.creationTimestamp")
-
+	createTime := common.FilterStr(KubernetesNS, "metadata.creationTimestamp")
 	timer, err := time.Parse(time.RFC3339, createTime)
-	DeploymentModel.CreateAt = timer
-	Name := common.FilterStr(KubernetesDEPLOY, "metadata.name")
-	DeploymentModel.Name = Name
-	DeploymentModel.Namespace = namespaceName
-	DeploymentModel.ClusterName = clusterName
-	DeploymentModel.WorkspaceName = workspaceName
-	Replicas := common.FilterStr(KubernetesDEPLOY, "status.replicas")
-	REPLICA.Replicas = StringToInt(Replicas)
-	ReadyReplicas := common.FilterStr(KubernetesDEPLOY, "status.readyReplicas")
-	REPLICA.ReadyReplicas = StringToInt(ReadyReplicas)
-	UpdatedReplicas := common.FilterStr(KubernetesDEPLOY, "status.updatedReplicas")
-	REPLICA.UpdatedReplicas = StringToInt(UpdatedReplicas)
-	AvailableReplicas := common.FilterStr(KubernetesDEPLOY, "status.availableReplicas")
-	REPLICA.AvailableReplicas = StringToInt(AvailableReplicas)
-	UnavailableReplicas := common.FilterStr(KubernetesDEPLOY, "status.unavailableReplicas")
-	REPLICA.UnavailableReplicas = StringToInt(UnavailableReplicas)
-	DeploymentModel.Replica = REPLICA
-	strategyType := common.Finding(KubernetesDEPLOY, "spec.strategy", "type")
-	strategyMaxUnavailable := common.Finding(KubernetesDEPLOY, "spec.strategy", "MaxUnavailable")
-	strategyMaxSurge := common.Finding(KubernetesDEPLOY, "spec.strategy", "MaxSurge")
-	STRATEGY.Type = strategyType
-	STRATEGY.MaxUnavailable = strategyMaxUnavailable
-	STRATEGY.MaxSurge = strategyMaxSurge
-	DeploymentModel.Strategy = STRATEGY
-	containers := common.FilterStr(KubernetesDEPLOY, "spec.template.spec.containers")
-	containersData := []model.CONTAINER{}
-	err4 := json.Unmarshal([]byte(containers), &containersData)
-	if err4 != nil {
-		panic(err4)
-	}
-
-	DeploymentModel.Containers = containersData
-	updateTime := common.Finding(KubernetesDEPLOY, "status.conditions", "lastUpdateTime")
-	timer2, err := time.Parse(time.RFC3339, updateTime)
-	DeploymentModel.UpdateAt = timer2
-	fmt.Printf("[#]updateTime : %+v\n", updateTime)
-
-	return c.JSON(http.StatusOK, echo.Map{"data": DeploymentModel})
+	projectModel.CreateAt = timer
+	return c.JSON(http.StatusOK, echo.Map{"data": projectModel})
 }
