+++ gmc_database_api_server/app/routes/routes.go
@@ -33,8 +33,8 @@ func GEdgeRoute(e *echo.Echo) {
 
 	r0 := e.Group("/test/v1")
 	// r0.GET("/cluster", api.Cluster)
-	r0.GET("/getCluster", api.Get_Cluster)
-	r0.GET("/getProject", api.Get_Project)
+	r0.GET("/getCluster/:name", api.Get_Cluster)
+	r0.GET("/getProject/:name", api.Get_Project)
 	r0.GET("/getDeployment/:name", api.Get_Deployment)
 	r0.GET("/gethttp", api.Get_http)
